Typecasting on #define

+++ misc/test3.c
@@ -19,7 +19,7 @@ int (*tests[])(void) = {
 	test2,
 	test3
 };
-#define TOTAL_TESTS ((int)sizeof(tests)/sizeof(int (*)(void)))
+#define TOTAL_TESTS ((long unsigned int)(sizeof(tests)/sizeof(int (*)(void))))
 
 int main(int argc, char *argv[])
 {
@@ -28,8 +28,8 @@ int main(int argc, char *argv[])
 	if (argc == 1) {
 		do_log(write_log, LOG_NAME, "Running File Operation Tests...\n");
 		passed = run_tests(tests);
-		printf("Tests passed: %d/%u\n", passed, TOTAL_TESTS);
-		do_log(append_log, LOG_NAME, "Tests passed: %d/%u\n", passed, TOTAL_TESTS);
+		printf("Tests passed: %d/%lu\n", passed, TOTAL_TESTS);
+		do_log(append_log, LOG_NAME, "Tests passed: %d/%lu\n", passed, TOTAL_TESTS);
 		if (passed == 0) {
 			printf("All tests failed!\n");
 		} else {
